Make the site title in the header link to the homepage
The site title is now a hyperlink to the homepage of the site to make navigation
using only the mouse possible.

This change also removes the concept of a per-page configurable header and
footer, because those must remain the same across the entire site if we're just
going to be reloading the content as the user browses.